Linear-memory storage modes, memory64, and component-model lift fixes

+++ Wacs.Core/Wacs.Core/Instructions/Table.cs
@@ -64,8 +64,11 @@ namespace Wacs.Core.Instructions
                 $"Instruction table.get failed. Wrong type on stack.");
             //7.
             long i = context.OpStack.PopAddr();
-            //8.
-            if (i >= tab.Elements.Count || (i > (long)int.MaxValue))
+            //8. Unsigned bound: table64 indices arrive as u64
+            // bit-patterns through PopAddr; a negative-as-signed i
+            // casts to a huge ulong that always fails the bound.
+            if ((ulong)i >= (ulong)tab.Elements.Count
+                || i > (long)int.MaxValue)
             {
                 throw new TrapException("Trap in table.get");
             }
@@ -130,8 +133,8 @@ namespace Wacs.Core.Instructions
                 $"Instruction table.set found incorrect type on top of the Stack");
             //9.
             long i = context.OpStack.PopAddr();
-            //10.
-            if (i >= tab.Elements.Count)
+            //10. Unsigned bound for table64 indices.
+            if ((ulong)i >= (ulong)tab.Elements.Count)
             {
                 throw new TrapException("table.set index exceeds table size.");
             }
